loc_bot/requirements.txt
python-telegram-bot[rate-limiter]>=21.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
pydantic>=2.0
